requirements.txt
python-dotenv>=0.20.0
sagemaker==2.140.0
discord-py-interactions==5.8.0
requests
pillow